Structural improvement & Date sorting
Some source design changes improving the cohesion and reducing the program coupling. 

Also, the program now sort the trades by date and saves the data in a matrix before placing it in the workbook. The date and hour values are now saved in the same cell as date format reducing the visual noise. 

The Bitcoin values in the values table now shows as Bitcoin format.

In general, the program is better optimized, unnecessary modules were deleted and part of the code was redesigned.